requirements.txt
apache-airflow==2.6.3
apache-airflow-providers-postgres==5.1.0
